+++ packages/file_picker_android/lib/src/android_platform_file.dart
@@ -88,13 +88,13 @@ base class AndroidPlatformFile extends PlatformFile {
   }
 
   @override
-  Future<int> length() async {
+  Future<int?> length() async {
     final len = _bytesLength;
     if (len != null && len > 0) return len;
     try {
       return await xFile.length();
     } catch (_) {
-      return 0;
+      return null;
     }
   }